Add tested files.
