--- h.c
@@ -1,1 +1,0 @@
-as
