+++ helpers.c
