--- helpers.h
@@ -1,10 +1,0 @@
-#ifndef HELPERS_H
-#define HELPERS_H
-
-int _putchar(char c);
-void _puts_mod(char *str);
-unsigned int _strlen(char *s);
-int _atoi(char *s);
-void print_number(int n);
-
-#endif /* HELPERS_H */
