--- t_function.c
@@ -1,10 +1,0 @@
-#include "holberton.h"
-
-/**
- * fucntion -
- * @x:
- * Return:
- */
-void function(void x)
-{
-}
