--- t_header.c
@@ -1,14 +1,0 @@
-#ifndef HEADER_H
-#define HEADER_H
-
-/**
- * struct nameStruc -
- *
- * @x:
- */
-typedef struct nameStruc
-{
-	void x;
-} nameStruc;
-
-#endif /* HEADER_H */
